reduced the code
