+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/TestCaseGeneratedByLLM.kt
@@ -1,5 +1,8 @@
 package org.jetbrains.research.testspark.core.test.data
 
+import org.jetbrains.research.testspark.core.test.data.strategies.PrintTestBodyStrategy
+import org.jetbrains.research.testspark.core.utils.Language
+
 /**
  *
  * Represents a test case generated by LLM.
@@ -11,6 +14,7 @@ data class TestCaseGeneratedByLLM(
     var expectedException: String = "",
     var throwsException: String = "",
     var lines: MutableList<TestLine> = mutableListOf(),
+    val language: Language
 ) {
 
     /**
@@ -104,31 +108,10 @@ data class TestCaseGeneratedByLLM(
      * @return a string containing the body of test case
      */
     private fun printTestBody(testInitiatedText: String): String {
-        var testFullText = testInitiatedText
-
-        // start writing the test signature
-        testFullText += "\n\tpublic void $name() "
-
-        // add throws exception if exists
-        if (throwsException.isNotBlank()) {
-            testFullText += "throws $throwsException"
-        }
-
-        // start writing the test lines
-        testFullText += "{\n"
-
-        // write each line
-        lines.forEach { line ->
-            testFullText += when (line.type) {
-                TestLineType.BREAK -> "\t\t\n"
-                else -> "\t\t${line.text}\n"
-            }
-        }
-
-        // close test case
-        testFullText += "\t}\n"
-
-        return testFullText
+       return when (language) {
+           Language.Kotlin -> PrintTestBodyStrategy.printKotlinTestBody(testInitiatedText, lines, throwsException, name)
+           Language.Java -> PrintTestBodyStrategy.printJavaTestBody(testInitiatedText, lines, throwsException, name)
+       }
     }
 
     /**
