+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/TestSuiteGeneratedByLLM.kt
@@ -1,5 +1,6 @@
 package org.jetbrains.research.testspark.core.test.data
 
+
 /**
  * Represents a test suite generated by LLM.
  *
