+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/dependencies/TestCompilationDependencies.kt
@@ -6,7 +6,7 @@ import org.jetbrains.research.testspark.core.data.JarLibraryDescriptor
  * The class represents a list of dependencies required for java test compilation.
  * The libraries listed are used during test suite/test case compilation.
  */
-class JavaTestCompilationDependencies {
+class TestCompilationDependencies {
     companion object {
         fun getJarDescriptors() = listOf(
             JarLibraryDescriptor(
