+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/kotlin/KotlinJUnitTestSuiteParser.kt
@@ -4,6 +4,7 @@ import org.jetbrains.research.testspark.core.data.JUnitVersion
 import org.jetbrains.research.testspark.core.test.TestSuiteParser
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.core.test.strategies.JUnitTestSuiteParserStrategy
+import org.jetbrains.research.testspark.core.utils.Language
 
 class KotlinJUnitTestSuiteParser(
     private val packageName: String,
@@ -17,6 +18,7 @@ class KotlinJUnitTestSuiteParser(
             importPattern,
             packageName,
             testNamePattern = "fun",
+            Language.Kotlin
         )
     }
 }
