+++ src/main/kotlin/org/jetbrains/research/testspark/services/kotlin/KotlinTestCaseDisplayService.kt
@@ -32,8 +32,8 @@ import org.jetbrains.research.testspark.core.data.Report
 import org.jetbrains.research.testspark.core.data.TestCase
 import org.jetbrains.research.testspark.core.utils.Language
 import org.jetbrains.research.testspark.data.UIContext
-import org.jetbrains.research.testspark.display.kotlin.KotlinTopButtonsPanelFactory
-import org.jetbrains.research.testspark.display.kotlin.KotlinTestCasePanelFactory
+import org.jetbrains.research.testspark.display.TestCasePanelFactory
+import org.jetbrains.research.testspark.display.TopButtonsPanelFactory
 import org.jetbrains.research.testspark.helpers.ReportHelper
 import org.jetbrains.research.testspark.helpers.kotlin.KotlinClassBuilderHelper
 import org.jetbrains.research.testspark.kotlin.KotlinPsiClassWrapper
@@ -64,7 +64,7 @@ class KotlinTestCaseDisplayService(private val project: Project) : TestCaseDispl
 
     private var mainPanel: JPanel = JPanel()
 
-    private val topButtonsPanelFactory = KotlinTopButtonsPanelFactory(project)
+    private val topButtonsPanelFactory = TopButtonsPanelFactory(project, Language.Kotlin)
 
     private var applyButton: JButton = JButton(PluginLabelsBundle.get("applyButton"))
 
@@ -123,7 +123,7 @@ class KotlinTestCaseDisplayService(private val project: Project) : TestCaseDispl
         addSeparator()
 
         // TestCasePanelFactories array
-        val testCasePanelFactories = arrayListOf<KotlinTestCasePanelFactory>()
+        val testCasePanelFactories = arrayListOf<TestCasePanelFactory>()
 
         report.testCaseList.values.forEach {
             val testCase = it
@@ -148,7 +148,7 @@ class KotlinTestCaseDisplayService(private val project: Project) : TestCaseDispl
             testCasePanel.add(checkbox, BorderLayout.WEST)
 
             val testCasePanelFactory =
-                KotlinTestCasePanelFactory(project, language, testCase, editor, checkbox, uiContext, report)
+                TestCasePanelFactory(project, language, testCase, editor, checkbox, uiContext, report)
             testCasePanel.add(testCasePanelFactory.getUpperPanel(), BorderLayout.NORTH)
             testCasePanel.add(testCasePanelFactory.getMiddlePanel(), BorderLayout.CENTER)
             testCasePanel.add(testCasePanelFactory.getBottomPanel(), BorderLayout.SOUTH)
